Doc improvements.
