+++ doc/connecting-to-data/catalog-functions.md
@@ -1,8 +1,8 @@
 A Catalog Function is a parameterized service where the user supplies the parameters and gets back some result.  The Type column in the table below indicates the `"type"` property to use in the [Initialization File](../customizing/initialization-files.md).
 
-| Name | Type | Description |
-|------|------|-------------|
-| Web Processing Service (WPS) | `wps` | |
-| Places Like Me | `places-like-me-function` | |
-| Spatial Detailing | `spatial-detailing-function` | |
-| Why Am I Special | `why-am-i-special-function` | |
+| Name | Type |
+|------|------|
+| [Web Processing Service (WPS)](catalog-type-details/wps.md) | `wps` |
+| [Places Like Me](catalog-type-details/places-like-me-function.md) | `places-like-me-function` |
+| [Spatial Detailing](catalog-type-details/spatial-detailing-function.md) | `spatial-detailing-function` |
+| [Why Am I Special](catalog-type-details/why-am-i-special-function.md) | `why-am-i-special-function` |
