+++ doc/connecting-to-data/catalog-groups.md
@@ -1,16 +1,16 @@
 A Catalog Group is a folder in the TerriaJS catalog that contains [Catalog Items](catalog-items.md), [Catalog Functions](catalog-functions.md), and other groups.  The Type column in the table below indicates the `"type"` property to use in the [Initialization File](../customizing/initialization-files.md).
 
-| Name | Type | Description |
-|------|------|-------------|
-| ABS ITT | `abs-itt-dataset-list` | |
-| CKAN | `ckan` | |
-| Catalog Service for the Web (CSW) | `csw` | |
-| ArcGIS Server | `esri-group` | |
-| ArcGIS FeatureServer | `esri-featureServer-group` | |
-| ArcGIS MapServer | `esri-mapServer-group` | |
-| Group (Manual) | `group` | |
-| Socrata | `socrata` | |
-| Web Feature Service (WFS) | `wfs-getCapabilities` | |
-| WFS Features | `wfs-features-group` | |
-| Web Map Service (WMS) | `wms-getCapabilities` | |
-| Web Processing Service (WPS) | `wps-getCapabilities` | |
+| Name | Type |
+|------|------|
+| [ABS ITT](catalog-type-details/abs-itt-dataset-list.md) | `abs-itt-dataset-list` |
+| [CKAN](catalog-type-details/ckan.md) | `ckan` |
+| [Catalog Service for the Web (CSW)](catalog-type-details/csw.md) | `csw` |
+| [ArcGIS Server](catalog-type-details/esri-group.md) | `esri-group` |
+| [ArcGIS FeatureServer](catalog-type-details/esri-featureServer-group.md) | `esri-featureServer-group` |
+| [ArcGIS MapServer](catalog-type-details/esri-mapServer-group.md) | `esri-mapServer-group` |
+| [Group (Manual)](catalog-type-details/group.md) | `group` |
+| [Socrata](catalog-type-details/socrata.md) | `socrata` |
+| [Web Feature Service (WFS)](catalog-type-details/wfs-getCapabilities.md) | `wfs-getCapabilities` |
+| [WFS Features](catalog-type-details/wfs-features-group.md) | `wfs-features-group` |
+| [Web Map Service (WMS)](catalog-type-details/wms-getCapabilities.md) | `wms-getCapabilities` |
+| [Web Processing Service (WPS)](catalog-type-details/wps-getCapabilities.md) | `wps-getCapabilities` |
