+++ doc/mkdocs.yml
@@ -20,25 +20,41 @@ pages:
     - Catalog Items: 'connecting-to-data/catalog-items.md'
     - Catalog Functions: 'connecting-to-data/catalog-functions.md'
     - Catalog Type Details:
-        - ABS ITT: 'connecting-to-data/catalog-type-details/abs-itt.md'
-        - Bing Maps: 'connecting-to-data/catalog-type-details/bing-maps.md'
-        - CKAN Resource: 'connecting-to-data/catalog-type-details/ckan-resource.md'
-        # - Composite: 'connecting-to-data/catalog-type-details/composite.md'
-        - Comma-Separated Values (CSV): 'connecting-to-data/catalog-type-details/csv.md'
-        - Cesium Terrain: 'connecting-to-data/catalog-type-details/cesium-terrain.md'
-        - Cesium Language (CZML): 'connecting-to-data/catalog-type-details/czml.md'
-        - ArcGIS FeatureServer: 'connecting-to-data/catalog-type-details/esri-featureServer.md'
-        - ArcGIS MapServer: 'connecting-to-data/catalog-type-details/esri-mapServer.md'
-        - GeoJSON: 'connecting-to-data/catalog-type-details/geojson.md'
-        - GPX: 'connecting-to-data/catalog-type-details/gpx.md'
-        - Keyhole Markup Language (KML): 'connecting-to-data/catalog-type-details/kml.md'
-        - Format Conversion Service (based on OGR): 'connecting-to-data/catalog-type-details/ogr.md'
-        - OpenStreetMap: 'connecting-to-data/catalog-type-details/open-street-map.md'
-        - SDMX-JSON: 'connecting-to-data/catalog-type-details/sdmx-json.md'
-        - Sensor Observation Service (SOS): 'connecting-to-data/catalog-type-details/sos.md'
-        - URL Template: 'connecting-to-data/catalog-type-details/url-template.md'
-        - Web Feature Service (WFS): 'connecting-to-data/catalog-type-details/wfs.md'
-        - Web Map Service (WMS): 'connecting-to-data/catalog-type-details/wms.md'
+        - ABS ITT (group): 'connecting-to-data/catalog-type-details/abs-itt-dataset-list.md'
+        - ABS ITT (item): 'connecting-to-data/catalog-type-details/abs-itt.md'
+        - ArcGIS FeatureServer (group): 'connecting-to-data/catalog-type-details/esri-featureServer-group.md'
+        - ArcGIS FeatureServer (item): 'connecting-to-data/catalog-type-details/esri-featureServer.md'
+        - ArcGIS MapServer (group): 'connecting-to-data/catalog-type-details/esri-mapServer-group.md'
+        - ArcGIS MapServer (item): 'connecting-to-data/catalog-type-details/esri-mapServer.md'
+        - ArcGIS Server (group): 'connecting-to-data/catalog-type-details/esri-group.md'
+        - Bing Maps (item): 'connecting-to-data/catalog-type-details/bing-maps.md'
+        - Catalog Service for the Web (CSW) (group): 'connecting-to-data/catalog-type-details/csw.md'
+        - Cesium Language (CZML) (item): 'connecting-to-data/catalog-type-details/czml.md'
+        - Cesium Terrain (item): 'connecting-to-data/catalog-type-details/cesium-terrain.md'
+        - CKAN (group): 'connecting-to-data/catalog-type-details/ckan.md'
+        - CKAN Resource (item): 'connecting-to-data/catalog-type-details/ckan-resource.md'
+        - Comma-Separated Values (CSV) (item): 'connecting-to-data/catalog-type-details/csv.md'
+        - Composite (item): 'connecting-to-data/catalog-type-details/composite.md'
+        - Format Conversion Service (based on OGR) (item): 'connecting-to-data/catalog-type-details/ogr.md'
+        - GeoJSON (item): 'connecting-to-data/catalog-type-details/geojson.md'
+        - GPX (item): 'connecting-to-data/catalog-type-details/gpx.md'
+        - Group (Manual) (group): 'connecting-to-data/catalog-type-details/group.md'
+        - Keyhole Markup Language (KML) (item): 'connecting-to-data/catalog-type-details/kml.md'
+        - OpenStreetMap (item): 'connecting-to-data/catalog-type-details/open-street-map.md'
+        - Places Like Me (function): 'connecting-to-data/catalog-type-details/places-like-me-function.md'
+        - SDMX-JSON (item): 'connecting-to-data/catalog-type-details/sdmx-json.md'
+        - Sensor Observation Service (SOS) (item): 'connecting-to-data/catalog-type-details/sos.md'
+        - Socrata (group): 'connecting-to-data/catalog-type-details/socrata.md'
+        - Spatial Detailing (function): 'connecting-to-data/catalog-type-details/spatial-detailing-function.md'
+        - URL Template (item): 'connecting-to-data/catalog-type-details/url-template.md'
+        - Web Feature Service (WFS) (group): 'connecting-to-data/catalog-type-details/wfs-getCapabilities.md'
+        - Web Feature Service (WFS) (item): 'connecting-to-data/catalog-type-details/wfs.md'
+        - Web Map Service (WMS) (group): 'connecting-to-data/catalog-type-details/wms-getCapabilities.md'
+        - Web Map Service (WMS) (item): 'connecting-to-data/catalog-type-details/wms.md'
+        - Web Processing Service (WPS) (function): 'connecting-to-data/catalog-type-details/wps.md'
+        - Web Processing Service (WPS) (group): 'connecting-to-data/catalog-type-details/wps-getCapabilities.md'
+        - WFS Features (group): 'connecting-to-data/catalog-type-details/wfs-features-group.md'
+        - Why Am I Special (function): 'connecting-to-data/catalog-type-details/why-am-i-special-function.md'
 - Deploying:
     - Overview: 'deploying/README.md'
     - Deploying TerriaMap: 'deploying/deploying-terriamap.md'
