+++ gulpfile.js
@@ -49,6 +49,7 @@ gulp.task('make-schema', function() {
             './lib/Models/*CatalogItem.js',
             './lib/Models/*CatalogGroup.js',
             './lib/Models/*CatalogMember.js',
+            './lib/Models/*CatalogFunction.js',
             '!./lib/Models/addUserCatalogMember.js',
             '!./lib/Models/AsyncFunctionResultCatalogItem.js'
         ],
@@ -180,7 +181,7 @@ function runKarma(configFile, done) {
     });
 }
 
-gulp.task('user-guide', ['make-schema'], function() {
+gulp.task('user-guide', /*['make-schema'],*/ function() {
     var fse = require('fs-extra');
     var gutil = require('gulp-util');
     var path = require('path');
