+++ lib/Models/CatalogItem.js
@@ -63,13 +63,6 @@ var CatalogItem = function(terria) {
     this.rectangle = undefined;
 
     /**
-     * Gets or sets the URL of the legend for this data item, or undefined if this data item does not have a legend.
-     * This property is observable.
-     * @type {LegendUrl}
-     */
-    this.legendUrl = undefined;
-
-    /**
      * Gets or sets the URL of this data.  This property is observable.
      * @type {String}
      */
@@ -200,6 +193,11 @@ var CatalogItem = function(terria) {
 
     var evaluatingLegendUrl = false;
 
+    /**
+     * Gets or sets the URLs of the legends to show when this catalog item is enabled.
+     * @member {LegendUrl} legendUrls
+     * @memberOf CatalogItem.prototype
+     */
     knockout.defineProperty(this, 'legendUrls', {
         get: function() {
             if (!defined(this._legendUrls) || this._legendUrls.length === 0) {
@@ -216,6 +214,12 @@ var CatalogItem = function(terria) {
         }
     });
 
+    /**
+     * Gets or sets the URL of the legend to show when this catalog item is enabled.  If there is more than one
+     * legend URL, this property returns the first one.
+     * @member {LegendUrl} legendUrl
+     * @memberOf CatalogItem.prototype
+     */
     knockout.defineProperty(this, 'legendUrl', {
         get: function() {
             evaluatingLegendUrl = true;
