+++ lib/Models/CompositeCatalogItem.js
@@ -15,8 +15,6 @@ var overrideProperty = require('../Core/overrideProperty');
 var CatalogItem = require('./CatalogItem');
 var createCatalogMemberFromType = require('./createCatalogMemberFromType');
 
-var compositeType = 'composite';
-
 /**
  * A {@link CatalogItem} composed of multiple other catalog items.  When this item is enabled or shown, the composed items are
  * enabled or shown as well.  Other properties, including {@link CatalogItem#rectangle},
@@ -68,7 +66,7 @@ defineProperties(CompositeCatalogItem.prototype, {
      */
     type : {
         get : function() {
-            return compositeType;
+            return 'composite';
         }
     },
 
@@ -163,7 +161,7 @@ CompositeCatalogItem.defaultUpdaters.items = function(compositeCatalogItem, json
             if (!defined(item.type)) {
                 throw new RuntimeError('Each item must have a type.');
             }
-            if (item.type === compositeType) {
+            if (item.type === 'composite') {
                 throw new RuntimeError('Composites cannot include composite items.');
             }
             var existingItem = createCatalogMemberFromType(item.type, compositeCatalogItem.terria);
