+++ lib/Models/WebProcessingServiceCatalogFunction.js
@@ -26,6 +26,15 @@ var Reproject = require('../Map/Reproject');
 var when = require('terriajs-cesium/Source/ThirdParty/when');
 var xml2json = require('../ThirdParty/xml2json');
 
+/**
+ * A {@link CatalogFunction} that invokes a Web Processing Service (WPS) process.
+ *
+ * @alias WebProcessingServiceCatalogFunction
+ * @constructor
+ * @extends CatalogFunction
+ *
+ * @param {Terria} terria The Terria instance.
+ */
 function WebProcessingServiceCatalogFunction(terria) {
     CatalogFunction.call(this, terria);
