+++ lib/Models/registerAnalytics.js
@@ -2,14 +2,14 @@
 
 /*global require*/
 var createCatalogMemberFromType = require('./createCatalogMemberFromType');
-var PlacesLikeMeFunction = require('./PlacesLikeMeFunction');
-var SpatialDetailingFunction = require('./SpatialDetailingFunction');
-var WhyAmISpecialFunction = require('./WhyAmISpecialFunction');
+var PlacesLikeMeCatalogFunction = require('./PlacesLikeMeCatalogFunction');
+var SpatialDetailingCatalogFunction = require('./SpatialDetailingCatalogFunction');
+var WhyAmISpecialCatalogFunction = require('./WhyAmISpecialCatalogFunction');
 
 var registerAnalytics = function() {
-    createCatalogMemberFromType.register('places-like-me-function', PlacesLikeMeFunction);
-    createCatalogMemberFromType.register('spatial-detailing-function', SpatialDetailingFunction);
-    createCatalogMemberFromType.register('why-am-i-special-function', WhyAmISpecialFunction);
+    createCatalogMemberFromType.register('places-like-me-function', PlacesLikeMeCatalogFunction);
+    createCatalogMemberFromType.register('spatial-detailing-function', SpatialDetailingCatalogFunction);
+    createCatalogMemberFromType.register('why-am-i-special-function', WhyAmISpecialCatalogFunction);
 };
 
 module.exports = registerAnalytics;
